fix stack functions
