+++ cpu.h
@@ -32,7 +32,4 @@ uint8_t get_arg(int n);
 void push(uint8_t value);
 uint8_t pop();
 
-void print_cpu();
-void print_ram();
-
 #endif
