+++ debug.h
@@ -1,0 +1,13 @@
+//
+// Created by cst on 11/1/20.
+//
+
+#ifndef MOS6502_DEBUG_H
+#define MOS6502_DEBUG_H
+
+void print_cpu();
+void print_ram();
+void print_stack();
+void info_cpu();
+
+#endif //MOS6502_DEBUG_H
